requirements.txt
# Core dependencies
numpy>=1.21.0
opencv-python>=4.5.0
pillow>=8.0.0
tqdm>=4.60.0

# Google Drive integration (for future use)
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.4.0
google-auth>=2.0.0

# Image processing and analysis
imagehash>=4.2.0

# For testing/debugging (optional)
matplotlib>=3.5.0

# Optional: for better zip handling
py7zr>=0.16.0

# OpenAI API for captioning
openai>=1.0.0

# Environment variables for API keys
python-dotenv>=0.19.0
